Add throw exception to comment module and post module

+++ src/comment/comment.service.ts
@@ -1,16 +1,19 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectModel } from 'nestjs-typegoose';
 import { CommentModel } from '../model/comment.model';
-import { ReturnModelType } from '@typegoose/typegoose';
+import { mongoose, ReturnModelType } from '@typegoose/typegoose';
 import { PostModel } from 'src/model/post.model';
+import { PostService } from 'src/post/post.service';
 
 @Injectable()
 export class CommentService {
+  PostService: any;
   constructor(
     @InjectModel(CommentModel)
     private readonly commentModel: ReturnModelType<typeof CommentModel>,
     @InjectModel(PostModel)
-    private readonly postModel: ReturnModelType<typeof PostModel>
+    private readonly postModel: ReturnModelType<typeof PostModel>,
+    private readonly postService: PostService
   ) {}
 
   async find(filter: any): Promise<CommentModel[]> {
@@ -18,6 +21,13 @@ export class CommentService {
   }
 
   async findById(id: string): Promise<CommentModel> {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new BadRequestException('Cid is not valid');
+    }
+    const existCid = await this.exists(id).then((result) => { return result; });
+    if (!existCid) {
+      throw new BadRequestException('Cid does not exist');
+    }
     return this.commentModel.findById(id).exec();
   }
 
@@ -33,7 +43,11 @@ export class CommentService {
   }
 
   async create(comment: CommentModel): Promise<CommentModel> {
-    if (this.postModel.exists(comment.pid)) {
+    if (!mongoose.Types.ObjectId.isValid(comment.pid)) {
+      throw new BadRequestException('Pid is not valid');
+    }
+    const existPid = await this.postService.exists(comment.pid).then((result) => { return result; });
+    if (!existPid) {
       throw new BadRequestException('Pid does not exist');
     }
     const newComment = new this.commentModel(comment);
@@ -41,10 +55,24 @@ export class CommentService {
   }
 
   async edit(cid: string, newMsg: string): Promise<CommentModel> {
+    if (!mongoose.Types.ObjectId.isValid(cid)) {
+      throw new BadRequestException('Cid is not valid');
+    }
+    const existCid = await this.exists(cid).then((result) => { return result; });
+    if (!existCid) {
+      throw new BadRequestException('Cid does not exist');
+    }
     return this.commentModel.findByIdAndUpdate(cid, {msg: newMsg}, {new: true}).exec();
   }
 
   async delete(cid: string) {
+    if (!mongoose.Types.ObjectId.isValid(cid)) {
+      throw new BadRequestException('Cid is not valid');
+    }
+    const existCid = await this.exists(cid).then((result) => { return result; });
+    if (!existCid) {
+      throw new BadRequestException('Cid does not exist');
+    }
     this.commentModel.findByIdAndDelete(cid).exec();
   }
