+++ src/post/post.service.ts
@@ -1,7 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectModel } from 'nestjs-typegoose';
 import { PostModel } from '../model/post.model';
-import { ReturnModelType } from '@typegoose/typegoose';
+import { mongoose, ReturnModelType } from '@typegoose/typegoose';
 import { CommentModel } from 'src/model/comment.model';
 
 @Injectable()
@@ -18,6 +18,13 @@ export class PostService {
   }
 
   async findById(id: string): Promise<PostModel> {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new BadRequestException('Pid is not valid');
+    }
+    const existPid = await this.exists(id).then((result) => { return result; });
+    if (!existPid) {
+      throw new BadRequestException('Pid does not exist');
+    }
     return this.postModel.findById(id).exec();
   }
 
@@ -42,10 +49,24 @@ export class PostService {
   }
 
   async edit(pid: string, newTitle: string, newContent: string): Promise<PostModel> {
+    if (!mongoose.Types.ObjectId.isValid(pid)) {
+      throw new BadRequestException('Pid is not valid');
+    }
+    const existPid = await this.exists(pid).then((result) => { return result; });
+    if (!existPid) {
+      throw new BadRequestException('Pid does not exist');
+    }
     return this.postModel.findByIdAndUpdate(pid, {title: newTitle, content: newContent}, {new: true}).exec();
   }
 
   async delete(pid: string) {
+    if (!mongoose.Types.ObjectId.isValid(pid)) {
+      throw new BadRequestException('Pid is not valid');
+    }
+    const existPid = await this.exists(pid).then((result) => { return result; });
+    if (!existPid) {
+      throw new BadRequestException('Pid does not exist');
+    }
     this.postModel.findByIdAndDelete(pid).exec();
     this.commentModel.deleteMany({pid: pid}).exec();
   }
